Add command web page
Added webpage with command help information
