+++ TobysBot.Discord/Pages/Invite.cshtml
@@ -1,19 +1,2 @@
-@page
+@page "/invite"
 @model TobysBot.Discord.Pages.Invite
-
-@{
-    Layout = null;
-}
-
-<!DOCTYPE html>
-
-<html>
-<head>
-    <title></title>
-</head>
-<body>
-<div>
-    
-</div>
-</body>
-</html>
